Добавлены тесты для функционала включение/выключение уведомлений
