+++ .idea/compiler.xml
@@ -6,6 +6,7 @@
       <module name="65.app.myapplication" target="11" />
       <module name="65.application" target="11" />
       <module name="65.core" target="11" />
+      <module name="65.lib" target="11" />
       <module name="65.library" target="11" />
     </bytecodeTargetLevel>
   </component>
