+++ .idea/misc.xml
@@ -9,6 +9,13 @@
         <entry key="..\:/Users/toriz/Study65/app/src/main/res/layout/fragment_contact_list.xml" value="0.17391304347826086" />
         <entry key="..\:/Users/toriz/Study65/app/src/main/res/layout/item_contact.xml" value="0.17391304347826086" />
         <entry key="..\:/Users/toriz/Study65/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1" />
+        <entry key="..\:/Users/toriz/Study65/lib/src/main/res/drawable/divider.xml" value="0.3333333333333333" />
+        <entry key="..\:/Users/toriz/Study65/lib/src/main/res/drawable/ic_launcher_background.xml" value="0.3333333333333333" />
+        <entry key="..\:/Users/toriz/Study65/lib/src/main/res/drawable/ic_launcher_foreground.xml" value="0.3333333333333333" />
+        <entry key="..\:/Users/toriz/Study65/lib/src/main/res/layout/fragment_contact_details.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/toriz/Study65/lib/src/main/res/layout/fragment_contact_list.xml" value="0.1" />
+        <entry key="..\:/Users/toriz/Study65/lib/src/main/res/layout/item_contact.xml" value="0.1" />
+        <entry key="..\:/Users/toriz/Study65/lib/src/main/res/menu/search_menu.xml" value="0.1" />
         <entry key="..\:/Users/toriz/Study65/libraryy/src/main/res/layout/activity_main.xml" value="0.35260416666666666" />
         <entry key="..\:/Users/toriz/Study65/libraryy/src/main/res/menu/search_menu.xml" value="0.36666666666666664" />
       </map>
