+++ app/src/main/java/com/velagissellint/a65/di/app/AppComponent.kt
@@ -1,8 +1,8 @@
 package com.velagissellint.a65.di.app
 
 import android.app.Application
-import com.velagissellint.a65.all.containersDi.AppContainer
 import com.velagissellint.a65.App
+import com.velagissellint.a65.all.containersDi.AppContainer
 import com.velagissellint.a65.di.contactDetails.ContactDetailsModelComponent
 import com.velagissellint.a65.di.contactList.ContactListModelComponent
 import com.velagissellint.a65.di.getContactsCase.GetContactsModule
