+++ lib/build.gradle
@@ -19,10 +19,13 @@ android {
         correctErrorTypes true
     }
 
+
     compileSdkVersion 31
     defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 31
+        minSdk 21
+        targetSdk 31
+        versionCode 1
+        versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
