+++ lib/src/main/AndroidManifest.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.velagissellint.library">
+
+    package="com.velagissellint.a65">
 
     <uses-permission android:name="android.permission.READ_PHONE_STATE" />
     <uses-permission android:name="android.permission.INTERNET" />
@@ -10,11 +11,11 @@
 
     <application
         android:allowBackup="true"
+        android:theme="@style/Theme.65"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
-        android:supportsRtl="true"
-        android:theme="@style/Theme.65">
+        android:supportsRtl="true">
         <receiver android:name="com.velagissellint.a65.all.data.BroadcastReceiverForNotify" />
         <activity
             android:name="com.velagissellint.a65.all.presentation.MainActivity"
@@ -27,6 +28,4 @@
             </intent-filter>
         </activity>
     </application>
-
-
 </manifest>
