+++ lib/src/main/java/com/velagissellint/a65/all/presentation/contactList/DetailedInformationDiffCallback.kt
