+++ lib/src/main/res/drawable/divider.xml
