+++ lib/src/main/res/layout/fragment_contact_details.xml
