+++ lib/src/main/res/layout/fragment_contact_list.xml
