+++ lib/src/main/res/layout/item_contact.xml
