+++ lib/src/main/res/mipmap/ic_launcher.png
