+++ lib/src/main/res/values/strings.xml
@@ -1,0 +1,14 @@
+<resources xmlns:tools="http://schemas.android.com/tools">
+    <string name="app_name">65</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+
+    <string name="title_for_ContactDetailsFragment" tools:ignore="ExtraTranslation">Детали контакта</string>
+    <string name="title_for_ContactListFragment">Список контактов</string>
+    <string name="title_for_notify">65:</string>
+    <string name="content_text">\"Сегодня день рождения у \"</string>
+    <string name="id_for_channel">channelID</string>
+    <string name="name_for_channel">chanel</string>
+    <string name="no_access_to_contacts">Требуется доступ к контактам</string>
+    <string name="contact_not_found">Контакт не найден</string>
+</resources>
