--- library/src/main/java/com/velagissellint/a65/all/RepositoryImpl.kt
@@ -1,72 +1,0 @@
-package com.velagissellint.a65.all
-
-import android.app.AlarmManager
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
-import com.velagissellint.a65.DetailedInformationAboutContact
-import com.velagissellint.a65.useCase.broadcast.BroadcastRepositoryCase
-import com.velagissellint.a65.useCase.contactDetails.ContactDetailsRepositoryCase
-import com.velagissellint.a65.useCase.contactListCase.ContactListRepositoryCase
-import com.velagissellint.a65.all.data.BroadcastReceiverForNotify
-import com.velagissellint.a65.all.data.ContactsRepository
-import com.velagissellint.a65.putNextBirthday
-import io.reactivex.rxjava3.core.Single
-
-class RepositoryImpl(
-    private val context: Context,
-    private val contactsRepository: ContactsRepository
-) :
-    ContactListRepositoryCase, ContactDetailsRepositoryCase, BroadcastRepositoryCase {
-    override fun getContacts(): Single<List<DetailedInformationAboutContact>>? =
-        contactsRepository.getContactsSingle()
-
-    override fun getContact(id: Int): Single<DetailedInformationAboutContact>? =
-        contactsRepository.getContact(id)
-
-
-    override fun offReminder(id: Int) {
-        val intent = Intent(context, BroadcastReceiverForNotify::class.java)
-        val alarmIntent = id?.let {
-            PendingIntent.getBroadcast(
-                context,
-                it,
-                intent,
-                PendingIntent.FLAG_CANCEL_CURRENT
-            )
-        }
-        val alarmManager = context?.getSystemService(Context.ALARM_SERVICE) as? AlarmManager
-        alarmManager?.cancel(alarmIntent)
-        alarmIntent?.cancel()
-    }
-
-    override fun onReminder(
-        id: Int,
-        detailedInformationAboutContact: DetailedInformationAboutContact
-    ) {
-        val intent = Intent(context, BroadcastReceiverForNotify::class.java)
-        intent.putExtra(FULL_NAME, detailedInformationAboutContact.fullName)
-        intent.putExtra(CONTACT_BIRTHDAY, detailedInformationAboutContact.birthday)
-        intent.putExtra(CONTACT_ID, id)
-        val alarmIntent = id?.let {
-            PendingIntent.getBroadcast(
-                context,
-                it,
-                intent,
-                PendingIntent.FLAG_UPDATE_CURRENT
-            )
-        }
-        val alarmManager = context?.getSystemService(Context.ALARM_SERVICE) as? AlarmManager
-        alarmManager?.set(
-            AlarmManager.RTC_WAKEUP,
-            putNextBirthday(detailedInformationAboutContact.birthday),
-            alarmIntent
-        )
-    }
-
-    companion object {
-        private const val CONTACT_ID = "id"
-        private const val FULL_NAME = "FULL_NAME"
-        private const val CONTACT_BIRTHDAY = "contactBirthday"
-    }
-}
